+++ tests/integration/test_cmd_script.py
@@ -106,7 +106,7 @@ def test_install_license(teardown):
             "gamspy",
             "install",
             "license",
-            os.environ["NETWORK_LICENSE"],
+            os.environ["NETWORK_LICENSE_NON_ACADEMIC"],
         ],
         check=True,
     )